07/12/2020

+++ java/Search/src/us/lsi/alg/mochila/MochilaEdge.java
@@ -3,24 +3,24 @@ package us.lsi.alg.mochila;
 import us.lsi.graphs.virtual.ActionSimpleEdge;
 import us.lsi.mochila.datos.DatosMochila;
 
-public class MochilaEdge extends ActionSimpleEdge<MochilaVertex,Double> {
+public class MochilaEdge extends ActionSimpleEdge<MochilaVertex,Integer> {
 	
-	public static MochilaEdge of(MochilaVertex v1, MochilaVertex v2, Double a) {		
+	public static MochilaEdge of(MochilaVertex v1, MochilaVertex v2, Integer a) {		
 		return new MochilaEdge(v1, v2, a);
 	}
 
-	public Double a;
+	public Integer a;
 	
-	private MochilaEdge(MochilaVertex v1, MochilaVertex v2, Double a) {
+	private MochilaEdge(MochilaVertex v1, MochilaVertex v2, Integer a) {
 		super(v1, v2);
 		this.a = a;
-		super.weight = a*DatosMochila.getValor(v1.index);
+		super.weight = a*DatosMochila.getValor(v1.index).doubleValue();
 //		System.out.println(this);
 	}
 
 	@Override
 	public String toString() {
-		return String.format("(%d,%d,%.2f,%.2f)",this.source.index,this.target.index,a,this.getEdgeWeight());
+		return String.format("(%d,%d,%d,%.2f)",this.source.index,this.target.index,a,this.getEdgeWeight());
 	}
 	
 }
